Ohhh ehh well i gotta goo
You have to work through the word bernhard gave you go on and ehhh also you have a little problem in the TableCreator Project yes alsoooo no going to bed before done and no no brushing teeth you got this ma boy and also snap rahel

Good morning and have a nice day

+++ src/main/java/com/mic_cust/frontend/Data/Conv_Output.java
@@ -16,7 +16,7 @@ public class Conv_Output implements Serializable {
     public String CompName;
     public String Quotation;
     public String Title;
-    public int Maintenance;
+    public double Maintenance;
     public int Duration;
     //Other Selectables:
     public String[] TempScopeMod;
